+++ CHANGES.md
@@ -5,6 +5,7 @@ New release
 - minor bug fixes
 - add installation guide to README.md
 - add next sprint and prev sprint buttons
+- add meetings filter in Time entries report
 
 0.8.2 (13-05-2013)
 ===============